backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
aiosqlite>=0.19.0
pydantic-settings>=2.1.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
zhipuai>=2.0.1
python-multipart>=0.0.6
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pyjwt>=2.8.0
bcrypt>=4.0.0
